% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_AstralPlane.R
\docType{class}
\name{AstralPlane-class}
\alias{AstralPlane-class}
\title{Class "AstralPlane"
This class stores astral phylogenetic data}
\description{
Class "AstralPlane"
This class stores astral phylogenetic data
}
\section{Slots}{

\describe{
\item{\code{fileName}}{file name is here}

\item{\code{samples}}{sample names are here}

\item{\code{phylo}}{phylo object for tree structure}

\item{\code{nodeData}}{data for nodes}

\item{\code{edgeData}}{data for branches}

\item{\code{ConcordanceFactorData}}{associated data from concordance factor analysis}
}}

\keyword{classes}
