% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astralProjection.R
\name{astralProjection}
\alias{astralProjection}
\title{AstralProjection}
\usage{
astralProjection(
  astral.plane = NULL,
  local.posterior = TRUE,
  pie.plot = c("qscore", "genetree"),
  save.file = NULL,
  pie.colors = c("black", "grey", "white"),
  node.color.text = c("white"),
  node.color.bg = c("black"),
  tip.label.size = 1,
  pie.chart.size = 1
)
}
\arguments{
\item{astral.plane}{AstralPlane S4 object of data generated from AstralPlane function}

\item{local.posterior}{plot the local posterior support?}

\item{pie.plot}{select one to plot: 'qscore' the quartet support or 'genetree' proportion of gene trees that support a branch}

\item{save.file}{if you wish to save to file, put file name. Saves as PDF}

\item{pie.colors}{select three colors to plot your pie.plot}

\item{node.color.text}{if local.posterior = TRUE, select the color of posterior support text}

\item{node.color.bg}{if local.posterior = TRUE, select the color of posterior support background}

\item{tip.label.size}{size of the tip labels, passed to cex in plotting function}

\item{pie.chart.size}{size of pie chart, passed to edgelabel plotting function}
}
\value{
plots the phylogenetic tree and selected data associated with an AstralPlane object. Can optionally be saved to file as a PDF by giving save.file a file name.
}
\description{
Function for plotting data from the astralPlane format
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

astralProjection(astral.plane = astral.data,
                 local.posterior = TRUE,
                 pie.plot = "qscore",
                 save.file = "test-dataset.pdf",
                 pie.colors = c("purple", "blue", "green"),
                 node.color.text = c("white"),
                 node.color.bg = c("black"),
                 tip.label.size = 0.75,
                 pie.chart.size = 1)

}
