% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astralRunner.R
\name{astralRunner}
\alias{astralRunner}
\title{astralRunner}
\usage{
astralRunner(
  concat.genetree.folder = NULL,
  output.dir = NULL,
  overwrite = FALSE,
  astral.path = NULL,
  astral.t = 2,
  quiet = TRUE,
  multi.thread = TRUE,
  memory = "1g"
)
}
\arguments{
\item{concat.genetree.folder}{a folder of genetree files that are concatenated.}

\item{output.dir}{the output directory name for the astral file}

\item{overwrite}{overwrite = TRUE to overwrite existing files}

\item{astral.path}{the absolute file path to the ASTRAL-III jar file}

\item{astral.t}{the ASTRAL-III "t" parameter for different annotations, t = 2 is all annotation}

\item{quiet}{TRUE hides the screen output from astral}

\item{multi.thread}{TRUE to use Astral-MP multithreading}

\item{memory}{memory value to be passed to java. Should be in "Xg" format, X = an integer}
}
\value{
Astral is run using this function as a wrapper. The astral output is saved to file, and can optionally be read into R as an ape phylo object using load.tree = TRUE.
}
\description{
This function runs many astral analyses across a folder of concatenated gene trees
}
\examples{


}
