% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAstral.R
\name{batchAstral}
\alias{batchAstral}
\title{batchAstral}
\usage{
batchAstral(
  genetree.datasets = NULL,
  output.dir = NULL,
  overwrite = TRUE,
  taxa.remove = NULL,
  min.n.samples = 4,
  min.sample.prop = NULL,
  make.polytomy = TRUE,
  polytomy.limit = 0,
  astral.path = NULL,
  astral.t = 2,
  quiet = TRUE,
  multi.thread = TRUE,
  memory = "1g"
)
}
\arguments{
\item{genetree.datasets}{a folder with folders of genetrees to prepare for astral analyses}

\item{output.dir}{the save name for the output directory}

\item{overwrite}{whether to overwrite an existing dataset}

\item{taxa.remove}{species that you would like removed from each gene tree}

\item{min.n.samples}{the minimum number of samples to keep a gene tree}

\item{min.sample.prop}{the minimum proportion of samples to keep a gene tree}

\item{make.polytomy}{whether to collapse poorly supported nodes into polytomies}

\item{polytomy.limit}{if make.polytomy = TRUE, the threshold value for node collapsing}

\item{astral.path}{the absolute path to astral. Needed because astral needs it.}

\item{astral.t}{the t value from astral for different annotations.}

\item{quiet}{hides the screen output from astral if desired}

\item{multi.thread}{whether to use Astral-MP multithreading or not.}

\item{memory}{memory value to be passed to java. Should be in "Xg" format, X = an integer}
}
\value{
applies setupAstral and runAstral to a folder, that contains folders of gene trees. Facilitates running the preparation and ASTRAL-III analysis for many different datasets at once. Genetree files and the astral output are saved in the desired output directory that can be read in individually as an AstralPlane object and plooted using AstralProjection.
}
\description{
Function for running many astral analyses across many datasets. Runs setupAstral and runAstral when given a folder that contains folders of gene trees.
}
\examples{

genetree.folder = "file/path/to/folder/of/genetree/folders"
taxa.delete = c("species_one", "species_two")

batchAstral(genetree.datasets = genetree.folder,
            output.dir = "test-dataset",
            overwrite = TRUE,
            taxa.remove = taxa.delete,
            min.n.samples = 4,
            min.sample.prop = 0.1,
            make.polytomy = TRUE,
            polytomy.limit = 10,
            astral.path = astral.path,
            astral.t = 2,
            quiet = F,
            multi.thread = TRUE,
            memory = "8g")

}
