% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordanceFactors.R
\name{concordanceFactors}
\alias{concordanceFactors}
\title{concordanceRunner}
\usage{
concordanceFactors(
  species.tree = NULL,
  alignment = NULL,
  gene.trees = NULL,
  output.name = NULL,
  overwrite = TRUE,
  quiet = TRUE,
  threads = 1
)
}
\arguments{
\item{overwrite}{overwrite if file exists?}

\item{quiet}{hides the screen output from astral if desired}

\item{threads}{how many threads to use}

\item{alignment.dir}{a folder of alignment files that are concatenated.}

\item{species.tree.dir}{a folder of species trees (e.g. from astral) to assess concordance}

\item{genetree.dir}{a folder of genetree files that are concatenated.}

\item{output.dir}{the output directory name for the astral file}
}
\value{
IQTree concordance factors are run using this function as a wrapper. The CF output is saved to file, and can be read into R using the function "readConcordance".
}
\description{
This function runs many concordance factor analyses using IQTREE2
}
\examples{

concordanceFactors(species.tree = paste0(species.tree.dir, "/", a.files[i]),
                   alignment = paste0(alignment.dir, "/", align.file),
                   gene.trees = paste0(genetree.dir, "/", gene.tree),
                   output.name = paste0(output.dir, "/", dataset.name),
                   quiet = quiet, threads = threads)

}
