% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordanceRunner.R
\name{concordanceRunner}
\alias{concordanceRunner}
\title{concordanceRunner}
\usage{
concordanceRunner(
  alignment.dir = NULL,
  species.tree.dir = NULL,
  genetree.dir = NULL,
  output.dir = "concordance-factors",
  overwrite = TRUE,
  quiet = TRUE,
  threads = 1
)
}
\arguments{
\item{alignment.dir}{a folder of alignment files that are concatenated.}

\item{species.tree.dir}{a folder of species trees (e.g. from astral) to assess concordance}

\item{genetree.dir}{a folder of genetree files that are concatenated.}

\item{output.dir}{the output directory name for the astral file}

\item{overwrite}{overwrite if file exists?}

\item{quiet}{hides the screen output from astral if desired}

\item{threads}{how many threads to use}
}
\value{
IQTree concordance factors are run using this function as a wrapper. The CF output is saved to file, and can be read into R using the function "readConcordance".
}
\description{
This function runs many concordance factor analyses using IQTREE2
}
\examples{

paths
genetree.path = paste0(work.dir, "/filtered-genetrees-concatenated")
alignment.path = paste0(work.dir, "/filtered-alignments-concatenated")
output.path = "concordance-factors"
astral.tree.path = paste0(work.dir, "/", astral.dir)

concordanceRunner(alignment.dir = alignment.path,
                 species.tree.dir = astral.tree.path,
                 genetree.dir = genetree.path,
                 output.dir = "concordance-factors",
                 overwrite = TRUE,
                 quiet = TRUE,
                 threads  = 6)

}
