% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAstralPlaneCF.R
\name{createAstralPlaneCF}
\alias{createAstralPlaneCF}
\title{createAstralPlaneCF}
\usage{
createAstralPlaneCF(cf.file.name = NULL, outgroups = NULL, tip.length = 1)
}
\arguments{
\item{cf.file.name}{prefix file name of the file for concordance factors}

\item{outgroups}{a vector of outgroups to root the tree}

\item{tip.length}{arbitrary value for the terminal tip lengths, Astral does not compute this}

\item{astral.tree}{file path to your tree}
}
\value{
an S4 Object of class astralPlane. The object is slotted and contains: 1. Sample names, 2. phylogenetic tree rooted, 3. node.data table of ASTRAL-III node data, 4. edge.data table of ASTRAL-III branch data. All these make ASTRAL-III output readable and easy to use in other functions, or to plot using the astralProjection function.
}
\description{
Function for reading data into the astralPlane format
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
