% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readConcordance.R
\name{readConcordance}
\alias{readConcordance}
\title{readConcordance}
\usage{
readConcordance(file.name = NULL)
}
\arguments{
\item{file.name}{file name of the .cf.stat file excluding ".cf.stat"}
}
\value{
a data.frame with the concordance factor data
}
\description{
Function for reading concordance factor data into a simple table
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
