% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupAstral.R
\name{setupAstral}
\alias{setupAstral}
\title{setupAstral}
\usage{
setupAstral(
  genetree.folder = NULL,
  output.name = NULL,
  overwrite = TRUE,
  taxa.remove = NULL,
  min.n.samples = 4,
  min.sample.prop = NULL,
  make.polytomy = TRUE,
  polytomy.limit = 0
)
}
\arguments{
\item{genetree.folder}{a folder of genetrees to prepare for astral analyses}

\item{output.name}{the save name for your concatenated gene tree file}

\item{overwrite}{whether to overwrite an existing dataset}

\item{taxa.remove}{species that you would like removed from each gene tree}

\item{min.n.samples}{the minimum number of samples to keep a gene tree}

\item{min.sample.prop}{the minimum proportion of samples to keep a gene tree}

\item{make.polytomy}{whether to collapse poorly supported nodes into polytomies}

\item{polytomy.limit}{if make.polytomy = TRUE, the threshold value for node collapsing}
}
\value{
the function goes through each gene tree and applies the desired filters. It then writes all the gene trees that pass the filters to a single file for use in ASTRAL-III.
}
\description{
This function sets up an astral run from gene trees
}
\examples{

genetree.folder = "file/path/to/folder/of/genetrees"
taxa.delete = c("species_one", "species_two")
save.name = "test-dataset"

setupAstral(genetree.folder = genetree.folder,
            output.name = save.name,
            overwrite = TRUE,
            taxa.remove = taxa.delete,
            min.n.samples = 4,
            min.sample.prop = 0.1,
            make.polytomy = TRUE,
            polytomy.limit = 10)

}
