% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_astralPlane.R
\docType{class}
\name{astralPlane-class}
\alias{astralPlane-class}
\title{Class "astralPlane"
This class stores astral phylogenetic data}
\description{
Class "astralPlane"
This class stores astral phylogenetic data
}
\section{Slots}{

\describe{
\item{\code{samples}}{sample names are here}

\item{\code{phylo}}{phylo object for tree structure}

\item{\code{nodeData}}{newick tree string}

\item{\code{edgeData}}{associated data}
}}

\keyword{classes}
