% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astralRunner.R
\name{astralRunner}
\alias{astralRunner}
\title{astralRunner}
\usage{
astralRunner(
  concat.genetree.folder = NULL,
  output.dir = NULL,
  overwrite = TRUE,
  astral.path = NULL,
  astral.t = 2,
  quiet = TRUE,
  multi.thread = TRUE,
  memory = "1g"
)
}
\arguments{
\item{concat.genetree.folder}{a folder of genetree files that are concatenated.}

\item{output.dir}{the output directory name for the astral file}

\item{overwrite}{overwrite if file exists?}

\item{astral.path}{the absolute path to astral. Needed because astral needs it.}

\item{astral.t}{the t value from astral for different annotations. t = 2, all annotations is recommended for full functionality}

\item{quiet}{hides the screen output from astral if desired}

\item{multi.thread}{whether to use Astral-MP multithreading or not.}

\item{memory}{memory value to be passed to java. Should be in "Xg" format, X = an integer}
}
\value{
Astral is run using this function as a wrapper. The astral output is saved to file, and can optionally be read into R as an ape phylo object using load.tree = TRUE.
}
\description{
This function runs many astral analyses across a folder of concatenated gene trees
}
\examples{


}
