% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordanceFactors.R
\name{concordanceFactors}
\alias{concordanceFactors}
\title{concordanceFactors}
\usage{
concordanceFactors(
  species.tree = NULL,
  alignment = NULL,
  gene.trees = NULL,
  output.name = NULL,
  overwrite = FALSE,
  quiet = TRUE,
  threads = 1
)
}
\arguments{
\item{species.tree}{a species trees (e.g. from astral) to assess concordance}

\item{alignment}{a concatenated alignment file}

\item{gene.trees}{a folder of genetree files that are concatenated.}

\item{output.name}{the output name for the concordance files}

\item{overwrite}{overwrite = TRUE to overwrite existing files}

\item{quiet}{hides the screen output from astral if desired}

\item{threads}{how many threads to use}
}
\value{
IQ-TREE 2 concordance factors files
}
\description{
This function runs concordance factor analyses using IQTREE2
}
\examples{

concordanceFactors(species.tree = paste0(species.tree.dir, "/", a.files[i]),
                   alignment = paste0(alignment.dir, "/", align.file),
                   gene.trees = paste0(genetree.dir, "/", gene.tree),
                   output.name = paste0(output.dir, "/", dataset.name),
                   quiet = quiet, threads = threads)

}
